brain-tumor-app/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
Pillow>=10.3.0
numpy>=1.26.0
tensorflow>=2.16.0
opencv-python>=4.9.0
pydantic>=2.7.0